Update to better handle non-ASCII names
This fixes a problem with init_logging() which didn't open the log file for utf-8 encoding, and thus visit_file() raised an error when writing non-ASCII names to the log.
I also added code to handle a UnicodeEncodeError in sanitize_filename() and then URL encode the name (e.g. %41%42%43), but this may not be needed.  It could also make filenames that are too long.